% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redsb.R
\docType{data}
\name{redsb}
\alias{redsb}
\title{Data to show stunting prevalence in every district from an island}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 33 rows and 15 columns.
}
\usage{
redsb
}
\description{
Data to show stunting prevalence in every district from an island
}
\keyword{datasets}
